+++ lib/utils/format.ts
@@ -1,4 +1,4 @@
-import { PaymentType, Province } from '@lib/graphql/types';
+import { Province } from '@lib/graphql/types';
 
 /**
  * Format North American phone number by removing all non-numeric characters.
@@ -104,22 +104,4 @@ ${province ? `${city} ${province}` : city}
 ${country || ''}
 ${postalCode}
   `;
-};
-
-/**
- * Format database payment types to respective CSV column name
- * @param {PaymentType} paymentMethod payment method to be formatted
- * @returns {string} formatted payment method
- */
-export const formatPaymentType = (paymentMethod: PaymentType): string => {
-  const paymentTypeToString: Record<PaymentType, string> = {
-    MASTERCARD: 'MasterCard (Office)',
-    VISA: 'Visa (Office)',
-    CASH: 'Cash',
-    DEBIT: 'Interac - Debit (Office)',
-    SHOPIFY: 'Shopify',
-    ETRANSFER: 'E-Transfer',
-    CHEQUE: 'Cheque',
-  };
-  return paymentTypeToString[paymentMethod];
 };
