+++ prisma/schema.prisma
@@ -321,7 +321,7 @@ enum ApplicantStatus {
   ACTIVE
   INACTIVE
 
-   @@map("applicantstatus")
+  @@map("applicantstatus")
 }
 
 enum Gender {
@@ -329,7 +329,7 @@ enum Gender {
   FEMALE
   OTHER
 
-   @@map("gender")
+  @@map("gender")
 }
 
 enum PaymentType {
@@ -341,14 +341,14 @@ enum PaymentType {
   DEBIT
   SHOPIFY
 
-   @@map("paymenttype")
+  @@map("paymenttype")
 }
 
 enum PhysicianStatus {
   ACTIVE
   INACTIVE
 
-   @@map("physicianstatus")
+  @@map("physicianstatus")
 }
 
 enum Province {
@@ -366,7 +366,7 @@ enum Province {
   NT
   YT
 
-   @@map("province")
+  @@map("province")
 }
 
 enum Role {
@@ -374,7 +374,7 @@ enum Role {
   ACCOUNTING
   SECRETARY
 
-   @@map("role")
+  @@map("role")
 }
 
 enum ApplicationStatus {
@@ -383,7 +383,7 @@ enum ApplicationStatus {
   REJECTED
   COMPLETED
 
-   @@map("applicationstatus")
+  @@map("applicationstatus")
 }
 
 enum ReasonForReplacement {
@@ -391,21 +391,21 @@ enum ReasonForReplacement {
   STOLEN
   OTHER
 
-   @@map("reasonforreplacement")
+  @@map("reasonforreplacement")
 }
 
 enum PermitType {
   PERMANENT
   TEMPORARY
 
-   @@map("permittype")
+  @@map("permittype")
 }
 
 enum AccessibleConvertedVanLoadingMethod {
   SIDE_LOADING
   END_LOADING
 
-   @@map("accessibleconvertedvanloadingmethod")
+  @@map("accessibleconvertedvanloadingmethod")
 }
 
 enum ApplicationType {
@@ -413,7 +413,7 @@ enum ApplicationType {
   RENEWAL
   REPLACEMENT
 
-   @@map("applicationtype")
+  @@map("applicationtype")
 }
 
 enum MobilityAid {
@@ -423,7 +423,7 @@ enum MobilityAid {
   SCOOTER
   WALKER
 
-   @@map("mobilityaid")
+  @@map("mobilityaid")
 }
 
 enum PatientCondition {
@@ -432,7 +432,7 @@ enum PatientCondition {
   CANNOT_WALK_100M
   OTHER
 
-   @@map("patientcondition")
+  @@map("patientcondition")
 }
 
 enum RequiresWiderParkingSpaceReason {
@@ -440,12 +440,12 @@ enum RequiresWiderParkingSpaceReason {
   MEDICAL_REASONS
   OTHER
 
-   @@map("requireswiderparkingspacereason")
+  @@map("requireswiderparkingspacereason")
 }
 
 enum ShopifyPaymentStatus {
   PENDING
   RECEIVED
 
-   @@map("shopifypaymentstatus")
+  @@map("shopifypaymentstatus")
 }
